requirements.txt
smolagents
google-auth-oauthlib==1.2.0
google-auth==2.28.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
pydantic==2.8.2 
gradio